Adding check to make sure the orphan email is an object.

+++ app/Jobs/AdoptOrphanEmails.php
@@ -61,47 +61,52 @@ class AdoptOrphanEmails extends Job implements ShouldQueue
             $currentClientId = 0;
             $failedToProcess = false;
 
-            $emailRecordCount = Email::where( 'email_address' , $item->email_address )->count();
-            if ( $emailRecordCount > 0 ) {
-                $currentEmailId = Email::select( 'id' )->where( 'email_address' , $item->email_address )->pluck( 'id' )->first();
+            if ( is_object( $currentOrphan ) ) {
+                $email RecordCount = Email::where( 'email_address' , $item->email_address )->count();
+                if ( $emailRecordCount > 0 ) {
+                    $currentEmailId = Email::select( 'id' )->where( 'email_address' , $item->email_address )->pluck( 'id' )->first();
 
-                $clientRecordCount = EmailClientInstance::where( 'email_id' , $currentEmailId )->count();
-                if ( $clientRecordCount > 0 ) {
-                    $currentClientId = EmailClientInstance::select( 'client_id' )->where( 'email_id' , $currentEmailId )->pluck( 'client_id' )->first();
+                    $clientRecordCount = EmailClientInstance::where( 'email_id' , $currentEmailId )->count();
+                    if ( $clientRecordCount > 0 ) {
+                        $currentClientId = EmailClientInstance::select( 'client_id' )->where( 'email_id' , $currentEmailId )->pluck( 'client_id' )->first();
 
-                    try {
-                        DB::connection( 'reporting_data' )->statement("
-                            INSERT INTO email_actions
-                                ( email_id , client_id , esp_account_id , campaign_id , action_id , datetime , created_at , updated_at )    
-                            VALUES
-                                ( ? , ? , ? , ? , ? , ? , NOW() , NOW() )
-                            ON DUPLICATE KEY UPDATE
-                                email_id = email_id ,
-                                client_id = client_id ,
-                                esp_account_id = esp_account_id ,
-                                campaign_id = campaign_id ,
-                                action_id = action_id ,
-                                datetime = datetime ,
-                                created_at = created_at ,
-                                updated_at = NOW()" ,
-                            [
-                                $currentEmailId ,
-                                $currentClientId ,
-                                $currentOrphan->esp_account_id ,
-                                $currentOrphan->campaign_id ,
-                                $currentOrphan->action_id ,
-                                $currentOrphan->datetime
-                            ]
-                        );
+                        try {
+                            DB::connection( 'reporting_data' )->statement("
+                                INSERT INTO email_actions
+                                    ( email_id , client_id , esp_account_id , campaign_id , action_id , datetime , created_at , updated_at )    
+                                VALUES
+                                    ( ? , ? , ? , ? , ? , ? , NOW() , NOW() )
+                                ON DUPLICATE KEY UPDATE
+                                    email_id = email_id ,
+                                    client_id = client_id ,
+                                    esp_account_id = esp_account_id ,
+                                    campaign_id = campaign_id ,
+                                    action_id = action_id ,
+                                    datetime = datetime ,
+                                    created_at = created_at ,
+                                    updated_at = NOW()" ,
+                                [
+                                    $currentEmailId ,
+                                    $currentClientId ,
+                                    $currentOrphan->esp_account_id ,
+                                    $currentOrphan->campaign_id ,
+                                    $currentOrphan->action_id ,
+                                    $currentOrphan->datetime
+                                ]
+                            );
 
-                        $currentOrphan->delete();
+                            $currentOrphan->delete();
 
-                        $processed++;
-                    } catch ( Exception $e ) {
-                        Log::error( 'Failed to process email ' . $item->email_address );
-                        Log::error( $e->getMessage() );
-                        Log::error( $e->getTraceAsString() );
+                            $processed++;
+                        } catch ( Exception $e ) {
+                            Log::error( 'Failed to process email ' . $item->email_address );
+                            Log::error( $e->getMessage() );
+                            Log::error( $e->getTraceAsString() );
 
+                            $failedToProcess = true;
+                            $attempts++;
+                        }
+                    } else {
                         $failedToProcess = true;
                         $attempts++;
                     }
@@ -109,9 +114,6 @@ class AdoptOrphanEmails extends Job implements ShouldQueue
                     $failedToProcess = true;
                     $attempts++;
                 }
-            } else {
-                $failedToProcess = true;
-                $attempts++;
             }
 
             if ( $failedToProcess ) {
